MGDAPI-5809 Removed monitoring from codebase
