+++ pkg/config/ConfigReadWriter_moq.go
@@ -42,9 +42,6 @@ var _ ConfigReadWriter = &ConfigReadWriterMock{}
 //			ReadMarin3rFunc: func() (*Marin3r, error) {
 //				panic("mock out the ReadMarin3r method")
 //			},
-//			ReadMonitoringFunc: func() (*Monitoring, error) {
-//				panic("mock out the ReadMonitoring method")
-//			},
 //			ReadObservabilityFunc: func() (*Observability, error) {
 //				panic("mock out the ReadObservability method")
 //			},
@@ -97,9 +94,6 @@ type ConfigReadWriterMock struct {
 	// ReadMarin3rFunc mocks the ReadMarin3r method.
 	ReadMarin3rFunc func() (*Marin3r, error)
 
-	// ReadMonitoringFunc mocks the ReadMonitoring method.
-	ReadMonitoringFunc func() (*Monitoring, error)
-
 	// ReadObservabilityFunc mocks the ReadObservability method.
 	ReadObservabilityFunc func() (*Observability, error)
 
@@ -147,9 +141,6 @@ type ConfigReadWriterMock struct {
 		// ReadMarin3r holds details about calls to the ReadMarin3r method.
 		ReadMarin3r []struct {
 		}
-		// ReadMonitoring holds details about calls to the ReadMonitoring method.
-		ReadMonitoring []struct {
-		}
 		// ReadObservability holds details about calls to the ReadObservability method.
 		ReadObservability []struct {
 		}
@@ -186,7 +177,6 @@ type ConfigReadWriterMock struct {
 	lockReadGrafana                 sync.RWMutex
 	lockReadMCG                     sync.RWMutex
 	lockReadMarin3r                 sync.RWMutex
-	lockReadMonitoring              sync.RWMutex
 	lockReadObservability           sync.RWMutex
 	lockReadProduct                 sync.RWMutex
 	lockReadRHSSO                   sync.RWMutex
@@ -409,33 +399,6 @@ func (mock *ConfigReadWriterMock) ReadMarin3rCalls() []struct {
 	mock.lockReadMarin3r.RLock()
 	calls = mock.calls.ReadMarin3r
 	mock.lockReadMarin3r.RUnlock()
-	return calls
-}
-
-// ReadMonitoring calls ReadMonitoringFunc.
-func (mock *ConfigReadWriterMock) ReadMonitoring() (*Monitoring, error) {
-	if mock.ReadMonitoringFunc == nil {
-		panic("ConfigReadWriterMock.ReadMonitoringFunc: method is nil but ConfigReadWriter.ReadMonitoring was just called")
-	}
-	callInfo := struct {
-	}{}
-	mock.lockReadMonitoring.Lock()
-	mock.calls.ReadMonitoring = append(mock.calls.ReadMonitoring, callInfo)
-	mock.lockReadMonitoring.Unlock()
-	return mock.ReadMonitoringFunc()
-}
-
-// ReadMonitoringCalls gets all the calls that were made to ReadMonitoring.
-// Check the length with:
-//
-//	len(mockedConfigReadWriter.ReadMonitoringCalls())
-func (mock *ConfigReadWriterMock) ReadMonitoringCalls() []struct {
-} {
-	var calls []struct {
-	}
-	mock.lockReadMonitoring.RLock()
-	calls = mock.calls.ReadMonitoring
-	mock.lockReadMonitoring.RUnlock()
 	return calls
 }
