+++ pkg/config/manager.go
@@ -45,7 +45,6 @@ type ConfigReadWriter interface {
 	ReadThreeScale() (*ThreeScale, error)
 	ReadMarin3r() (*Marin3r, error)
 	GetOperatorNamespace() string
-	ReadMonitoring() (*Monitoring, error)
 	ReadProduct(product integreatlyv1alpha1.ProductName) (ConfigReadable, error)
 	ReadCloudResources() (*CloudResources, error)
 	ReadGrafana() (*Grafana, error)
@@ -97,8 +96,6 @@ func (m *Manager) ReadProduct(product integreatlyv1alpha1.ProductName) (ConfigRe
 		return m.ReadRHSSOUser()
 	case integreatlyv1alpha1.ProductCloudResources:
 		return m.ReadCloudResources()
-	case integreatlyv1alpha1.ProductMonitoring:
-		return m.ReadMonitoring()
 	case integreatlyv1alpha1.ProductMarin3r:
 		return m.ReadMarin3r()
 	case integreatlyv1alpha1.ProductGrafana:
@@ -150,14 +147,6 @@ func (m *Manager) ReadRHSSOUser() (*RHSSOUser, error) {
 		return nil, err
 	}
 	return NewRHSSOUser(config), nil
-}
-
-func (m *Manager) ReadMonitoring() (*Monitoring, error) {
-	config, err := m.readConfigForProduct(integreatlyv1alpha1.ProductMonitoring)
-	if err != nil {
-		return nil, err
-	}
-	return NewMonitoring(config), nil
 }
 
 func (m *Manager) ReadGrafana() (*Grafana, error) {
